Add uuid_unit (unsigned long int) as a typedef

+++ Code/Globals.h
@@ -26,7 +26,7 @@ void logConsole(const char file[], int line, const char* format, ...);
 
 
 typedef unsigned int uint;
-typedef unsigned int uuid_unit;
+typedef unsigned long int uuid_unit;
 
 enum update_status
 {
